requirements.txt
Django==4.2.7
daphne==4.0.0
channels==4.0.0
channels-redis==4.1.0
django-redis==5.3.0
psycopg2-binary==2.9.7
stripe==12.5.1
cryptography==46.0.1
django-two-factor-auth==1.18.1
pytest==8.4.2
pytest-django==4.11.1
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0
celery==5.3.4
redis==5.0.1
requests==2.32.3
Pillow==10.1.0
qrcode==8.2
django-phonenumber-field==8.3.0
django-formtools==2.5.1
openai==1.54.0
django-ratelimit==4.1.0
django-csp==3.8
django-security==1.1.0
geoip2==4.8.0
django-geoip==1.1.0
django-session-security==2.6.7
django-fernet-fields==0.1.4
celery[redis]==5.3.4
flower==2.0.1
google-api-python-client==2.110.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
django-allauth==0.61.1
requests-oauthlib==1.3.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
textblob==0.17.1
nltk==3.8.1
joblib==1.3.2
xgboost==2.0.0
tensorflow==2.15.0
transformers==4.35.2
torch==2.1.0
celery[redis]==5.3.4
flower==2.0.1
google-api-python-client==2.110.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
django-allauth==0.61.1
requests-oauthlib==1.3.1